asdf

+++ backend/product_api/models.py
@@ -8,7 +8,7 @@ class Product(models.Model):
     title = models.CharField(max_length=100)
     description = models.TextField()
     image = models.ImageField(upload_to='products/', null=True, blank=True)
-    price = models.FloatField(null=True, blank=True)
+    price = models.DecimalField(max_digits=10, decimal_places=2, null=True, blank=True)
     address = models.CharField(max_length=100, null=True, blank=True)
     user = models.ForeignKey(UserModel, on_delete=models.CASCADE)
     category = models.ForeignKey(Category, on_delete=models.CASCADE, blank=True, null=True)
